Move 'MainQueueDispatchDecorator' and its composition helpers to a separate file
